requirements.txt
flask==2.2.5
google_api_python_client==2.99.0
google_auth_oauthlib==1.1.0
requests==2.31.0
gunicorn==20.0.4
